added option to make views public by setting a boolean value
